code review and cleanup
